Update README.md
This repository is a collection of Git exercises. 

The exercises are designed for use when we are teaching Git courses. You should be able to use them as self-contained exercises that will allow you to keep your Git skills sharp.

Exercises starting with basic are entry-level - other exercises vary greatly in difficulty.